refactor: 뷰 홀더 생성 방식 변경
