+++ app/src/main/java/woowacourse/shopping/feature/cart/CartProductListAdapter.kt
@@ -1,9 +1,7 @@
 package woowacourse.shopping.feature.cart
 
-import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import woowacourse.shopping.databinding.ItemCartProductBinding
 import woowacourse.shopping.model.CartProductState
 
 class CartProductListAdapter(
@@ -14,10 +12,8 @@ class CartProductListAdapter(
 ) : RecyclerView.Adapter<CartProductViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CartProductViewHolder {
-        val inflater = LayoutInflater.from(parent.context)
-        val binding = ItemCartProductBinding.inflate(inflater, parent, false)
-        return CartProductViewHolder(
-            binding, onCartProductDeleteClick, onCountMinusClick, onCountPlusClick
+        return CartProductViewHolder.createInstance(
+            parent, onCartProductDeleteClick, onCountMinusClick, onCountPlusClick
         )
     }
