+++ app/src/main/java/woowacourse/shopping/feature/product/ProductViewHolder.kt
@@ -1,5 +1,7 @@
 package woowacourse.shopping.feature.product
 
+import android.view.LayoutInflater
+import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewbinding.ViewBinding
 import woowacourse.shopping.databinding.ItemProductBinding
@@ -11,5 +13,13 @@ class ProductViewHolder(binding: ViewBinding) : RecyclerView.ViewHolder(binding.
     fun bind(productState: ProductState, onClick: (ProductState) -> Unit) {
         binding.product = productState
         binding.root.setOnClickListener { onClick(productState) }
+    }
+
+    companion object {
+        fun createInstance(parent: ViewGroup): ProductViewHolder {
+            val inflater = LayoutInflater.from(parent.context)
+            val binding = ItemProductBinding.inflate(inflater, parent, false)
+            return ProductViewHolder(binding)
+        }
     }
 }
