+++ app/src/main/java/woowacourse/shopping/feature/product/recent/RecentProductItemViewHolder.kt
@@ -1,5 +1,7 @@
 package woowacourse.shopping.feature.product.recent
 
+import android.view.LayoutInflater
+import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewbinding.ViewBinding
 import woowacourse.shopping.databinding.ItemRecentBinding
@@ -10,5 +12,13 @@ class RecentProductItemViewHolder(binding: ViewBinding) : RecyclerView.ViewHolde
 
     fun bind(recentProductState: RecentProductState) {
         binding.recentProduct = recentProductState
+    }
+
+    companion object {
+        fun createInstance(parent: ViewGroup): RecentProductItemViewHolder {
+            val inflater = LayoutInflater.from(parent.context)
+            val binding = ItemRecentBinding.inflate(inflater, parent, false)
+            return RecentProductItemViewHolder(binding)
+        }
     }
 }
