+++ app/src/main/java/woowacourse/shopping/feature/product/recent/RecentProductWrapperViewHolder.kt
@@ -1,5 +1,7 @@
 package woowacourse.shopping.feature.product.recent
 
+import android.view.LayoutInflater
+import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewbinding.ViewBinding
 import woowacourse.shopping.databinding.ItemRecentProductListBinding
@@ -12,5 +14,13 @@ class RecentProductWrapperViewHolder(
     fun bind() {
         binding as ItemRecentProductListBinding
         binding.recentListRv.adapter = adapter
+    }
+
+    companion object {
+        fun createInstance(parent: ViewGroup, recentProductListAdapter: RecentProductListAdapter): RecentProductWrapperViewHolder {
+            val inflater = LayoutInflater.from(parent.context)
+            val binding = ItemRecentProductListBinding.inflate(inflater, parent, false)
+            return RecentProductWrapperViewHolder(binding, recentProductListAdapter)
+        }
     }
 }
