+++ app/src/main/res/drawable/bg_round_square_white.xml
@@ -4,9 +4,6 @@
     android:padding="10dp"
     android:shape="rectangle" >
     <solid android:color="#FFF" />
-    <stroke
-        android:width="1dp"
-        android:color="@color/gray"/>
     <corners
         android:radius="6dp" />
 </shape>
